+++ tests/unit/structures/test_structure.py
@@ -1,6 +1,7 @@
 import pytest
 
-from griptape.structures import Agent, Pipeline
+from griptape.structures import Agent, Pipeline, Structure
+from griptape.tasks import PromptTask
 
 
 class TestStructure:
@@ -16,3 +17,40 @@ class TestStructure:
             ValueError, match="Structure's output Task has no output. Run the Structure to generate output."
         ):
             assert agent.output
+
+    def test_conversation_mode_per_structure(self):
+        pipeline = Pipeline(
+            conversation_memory_strategy=Structure.ConversationMemoryStrategy.PER_STRUCTURE,
+            tasks=[PromptTask("1"), PromptTask("2")],
+        )
+
+        pipeline.run()
+        assert pipeline.conversation_memory is not None
+        assert len(pipeline.conversation_memory.runs) == 1
+        assert pipeline.conversation_memory.runs[0].input.value == "1"
+        assert pipeline.conversation_memory.runs[0].output.value == "mock output"
+
+    def test_conversation_mode_per_task(self):
+        pipeline = Pipeline(
+            conversation_memory_strategy=Structure.ConversationMemoryStrategy.PER_TASK,
+            tasks=[PromptTask("1"), PromptTask("2")],
+        )
+
+        pipeline.run()
+        assert pipeline.conversation_memory is not None
+        assert len(pipeline.conversation_memory.runs) == 2
+        assert pipeline.conversation_memory.runs[0].input.value == "1"
+        assert pipeline.conversation_memory.runs[0].output.value == "mock output"
+        assert pipeline.conversation_memory.runs[1].input.value == "2"
+        assert pipeline.conversation_memory.runs[1].output.value == "mock output"
+
+    def test_conversation_mode_per_task_no_memory(self):
+        pipeline = Pipeline(
+            conversation_memory_strategy=Structure.ConversationMemoryStrategy.PER_TASK,
+            tasks=[PromptTask(conversation_memory=None), PromptTask("2")],
+        )
+
+        pipeline.run()
+        assert pipeline.conversation_memory is not None
+        assert len(pipeline.conversation_memory.runs) == 1
+        assert pipeline.conversation_memory.runs[0].input.value == "2"